Components! (#47)
Convert progress bar to a component so it's nice and portable.
Create a component-displayer-component so we can pass the name of a step in and get the right component back. This way, we can make a workflow with a simple string list (or array), and have it automatically calculate now long the workflow is. This also drastically simplifies the code, which is always a good thing.

Re-arranging a workflow is as simple as re-arranging it's list.